transformer/requirements.txt
xmltodict==0.13.0
psycopg2-binary==2.9.9
azure-servicebus==7.11.4
python-gnupg==0.5.2
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
